requirements.txt
# --- API core ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# --- NLP / Embeddings / Indexing ---
sentence-transformers==3.0.1
torch==2.3.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
tqdm==4.66.4
scikit-learn==1.5.1

# --- Parsers / Loaders ---
pymupdf==1.24.9
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
chardet==5.2.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.3.0
